--- src/transaction/makeInputTemplate.js
@@ -1,7 +1,0 @@
-export default function makeInputTemplate(publicKeys = [], fulfills = null, fulfillment = null) {
-    return {
-        fulfillment,
-        fulfills,
-        'owners_before': publicKeys,
-    }
-}
